+++ packages/tools-react-native/src/index.ts
@@ -9,6 +9,7 @@ export {
   getMetroVersion,
   requireModuleFromMetro,
 } from "./metro.ts";
+export { mergeTransformerConfigs } from "./metro-utils.ts";
 export {
   expandPlatformExtensions,
   getAvailablePlatforms,