+++ scripts/dependencies.config.js
@@ -4,10 +4,29 @@ import * as fs from "node:fs";
 import * as path from "node:path";
 
 /**
+ * @typedef {{
+ *   scripts?: Record<string, string>;
+ *   dependencies?: Record<string, string>;
+ *   jest?: unknown;
+ * }} Manifest;
+ */
+
+/**
  * @returns {true}
  */
 function always() {
   return true;
+}
+
+/**
+ * @template {object} T
+ * @template {keyof T} K
+ * @param {T} obj
+ * @param {K} property
+ * @returns {obj is T & Pick<Required<T>, K>}
+ */
+function hasOwn(obj, property) {
+  return Object.hasOwn(obj, property);
 }
 
 /**
@@ -20,25 +39,24 @@ function lookForFile(filename) {
 
 /**
  * @param {string} cwd
- * @param {Record<string, unknown>} manifest
+ * @param {Manifest} manifest
  * @returns {boolean}
  */
 function needsJest(cwd, manifest) {
   return (
-    Object.hasOwn(manifest, "jest") ||
-    fs.existsSync(path.join(cwd, "jest.config.js"))
+    hasOwn(manifest, "jest") || fs.existsSync(path.join(cwd, "jest.config.js"))
   );
 }
 
 /**
  * @param {string} _cwd
- * @param {Record<string, Record<string, string>>} manifest
+ * @param {Manifest} manifest
  * @returns {boolean}
  */
 function needsLint(_cwd, manifest) {
   return (
-    Object.hasOwn(manifest, "scripts") &&
-    Object.hasOwn(manifest["scripts"], "lint") &&
+    hasOwn(manifest, "scripts") &&
+    hasOwn(manifest["scripts"], "lint") &&
     manifest["scripts"]["lint"].includes("rnx-kit-scripts lint")
   );
 }
@@ -54,7 +72,11 @@ const COMMON_DEPENDENCIES = /** @type {const} */ ([
   ["typescript", needsTypeScript],
 ]);
 
+/**
+ * @param {{ cwd: string; manifest: Manifest; }}
+ */
 export default function ({ cwd, manifest }) {
+  /** @type {Pick<Required<Manifest>, "dependencies"> | undefined} */
   let extensions = undefined;
 
   for (const [pkg, test] of COMMON_DEPENDENCIES) {