+++ packages/tools-react-native/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Change Log - @rnx-kit/tools-react-native
 
+## 2.3.7
+
+### Patch Changes
+
+- 9ee8e97: Bump to retry publishing
+
+## 2.3.6
+
+### Patch Changes
+
+- 0bf51d8: Add function to merge transformer configs to tools-react-native and use it in the cli
+
 ## 2.3.5
 
 ### Patch Changes